pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-watermark"
version = "2.2.0"
description = "基于EXIF信息的照片日期水印工具"
authors = [{name = "NJUSE25FALL"}]
dependencies = [
    "Pillow>=10.0.0",
    "piexif>=1.1.3",
    "click>=8.0.0",
    "tkinterdnd2>=0.3.0"
]